+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## Unreleased
 
+- `ZarrOutput` for writing simulation output to a Zarr store, implemented as a Zarr.jl extension
 - Documentation hero page and docs rendering [#1071](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/1071) [#1072](https://github.com/SpeedyWeather/SpeedyWeather.jl/pull/1072)
-
 - Bump julia-actions/cache and setup-julia v2 to v3 [#1073](ttps://github.com/SpeedyWeather/SpeedyWeather.jl/pull/1073)
 
 ## v0.19.0